+++ src/main/java/dev/lambdaurora/lambdynlights/mixin/lightsource/AbstractMinecartEntityMixin.java
@@ -9,69 +9,29 @@
 
 package dev.lambdaurora.lambdynlights.mixin.lightsource;
 
-import dev.lambdaurora.lambdynlights.DynamicLightSource;
-import dev.lambdaurora.lambdynlights.LambDynLights;
-import dev.lambdaurora.lambdynlights.api.DynamicLightHandlers;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.vehicle.AbstractMinecart;
-import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.Unique;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 /**
  * Adds the tick method for dynamic light source tracking in minecart entities.
  *
  * @author LambdAurora
- * @version 2.0.2
+ * @version 3.3.0
  * @since 1.3.2
  */
 @Mixin(AbstractMinecart.class)
-public abstract class AbstractMinecartEntityMixin extends Entity implements DynamicLightSource {
+public abstract class AbstractMinecartEntityMixin extends EntityMixin {
 	@Shadow
 	public abstract BlockState getDisplayBlockState();
 
-	@Unique
-	private int lambdynlights$luminance;
-
-	public AbstractMinecartEntityMixin(EntityType<?> type, Level level) {
-		super(type, level);
-	}
-
-	@Inject(method = "tick", at = @At("HEAD"))
-	private void onTick(CallbackInfo ci) {
-		// We do not want to update the entity on the server.
-		if (this.level().isClientSide()) {
-			if (this.isRemoved()) {
-				this.setDynamicLightEnabled(false);
-			} else {
-				if (!LambDynLights.get().config.getEntitiesLightSource().get() || !DynamicLightHandlers.canLightUp(this))
-					this.lambdynlights$luminance = 0;
-				else
-					this.dynamicLightTick();
-				LambDynLights.updateTracking(this);
-			}
-		}
-	}
-
 	@Override
 	public void dynamicLightTick() {
+		super.dynamicLightTick();
 		this.lambdynlights$luminance = Math.max(
-				Math.max(
-						this.isOnFire() ? 15 : 0,
-						this.getDisplayBlockState().getLightEmission()
-				),
-				DynamicLightHandlers.getLuminanceFrom(this)
+				this.lambdynlights$luminance,
+				this.getDisplayBlockState().getLightEmission()
 		);
-	}
-
-	@Override
-	public int getLuminance() {
-		return this.lambdynlights$luminance;
 	}
 }