+++ src/main/java/dev/lambdaurora/lambdynlights/mixin/lightsource/LivingEntityMixin.java
@@ -9,45 +9,24 @@
 
 package dev.lambdaurora.lambdynlights.mixin.lightsource;
 
-import dev.lambdaurora.lambdynlights.DynamicLightSource;
 import dev.lambdaurora.lambdynlights.LambDynLights;
 import dev.lambdaurora.lambdynlights.api.DynamicLightHandlers;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Unique;
 
 @Mixin(LivingEntity.class)
-public abstract class LivingEntityMixin extends Entity implements DynamicLightSource {
-	@Unique
-	protected int lambdynlights$luminance;
-
-	public LivingEntityMixin(EntityType<?> type, Level level) {
-		super(type, level);
-	}
-
+public abstract class LivingEntityMixin extends EntityMixin {
 	@Override
 	public void dynamicLightTick() {
-		if (!LambDynLights.get().config.getEntitiesLightSource().get() || !DynamicLightHandlers.canLightUp(this)) {
-			this.lambdynlights$luminance = 0;
-			return;
-		}
-
 		if (this.isOnFire() || this.isCurrentlyGlowing()) {
 			this.lambdynlights$luminance = 15;
 		} else {
 			this.lambdynlights$luminance = LambDynLights.getLivingEntityLuminanceFromItems((LivingEntity) (Object) this);
 		}
 
-		int luminance = DynamicLightHandlers.getLuminanceFrom(this);
+		int luminance = DynamicLightHandlers.getLuminanceFrom((Entity) (Object) this);
 		if (luminance > this.lambdynlights$luminance)
 			this.lambdynlights$luminance = luminance;
-	}
-
-	@Override
-	public int getLuminance() {
-		return this.lambdynlights$luminance;
 	}
 }