+++ src/main/java/dev/lambdaurora/lambdynlights/mixin/lightsource/PlayerEntityMixin.java
@@ -9,11 +9,7 @@
 
 package dev.lambdaurora.lambdynlights.mixin.lightsource;
 
-import dev.lambdaurora.lambdynlights.DynamicLightSource;
-import dev.lambdaurora.lambdynlights.LambDynLights;
-import dev.lambdaurora.lambdynlights.api.DynamicLightHandlers;
-import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.LivingEntity;
+import dev.lambdaurora.lambdynlights.engine.DynamicLightSource;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
@@ -21,34 +17,16 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 
 @Mixin(Player.class)
-public abstract class PlayerEntityMixin extends LivingEntity implements DynamicLightSource {
+public abstract class PlayerEntityMixin extends LivingEntityMixin implements DynamicLightSource {
 	@Shadow
 	public abstract boolean isSpectator();
 
 	@Unique
-	protected int lambdynlights$luminance;
-	@Unique
 	private Level lambdynlights$lastWorld;
-
-	protected PlayerEntityMixin(EntityType<? extends LivingEntity> type, Level level) {
-		super(type, level);
-	}
 
 	@Override
 	public void dynamicLightTick() {
-		if (!DynamicLightHandlers.canLightUp(this)) {
-			this.lambdynlights$luminance = 0;
-			return;
-		}
-
-		if (this.isOnFire() || this.isCurrentlyGlowing()) {
-			this.lambdynlights$luminance = 15;
-		} else {
-			this.lambdynlights$luminance = Math.max(
-					DynamicLightHandlers.getLuminanceFrom(this),
-					LambDynLights.getLivingEntityLuminanceFromItems(this)
-			);
-		}
+		super.dynamicLightTick();
 
 		if (this.isSpectator())
 			this.lambdynlights$luminance = 0;
@@ -57,10 +35,5 @@ public abstract class PlayerEntityMixin extends LivingEntity implements DynamicL
 			this.lambdynlights$lastWorld = this.level();
 			this.lambdynlights$luminance = 0;
 		}
-	}
-
-	@Override
-	public int getLuminance() {
-		return this.lambdynlights$luminance;
 	}
 }