+++ src/main/java/dev/lambdaurora/lambdynlights/mixin/lightsource/PrimedTntEntityMixin.java
@@ -9,11 +9,10 @@
 
 package dev.lambdaurora.lambdynlights.mixin.lightsource;
 
-import dev.lambdaurora.lambdynlights.DynamicLightSource;
 import dev.lambdaurora.lambdynlights.ExplosiveLightingMode;
 import dev.lambdaurora.lambdynlights.LambDynLights;
-import dev.lambdaurora.lambdynlights.api.DynamicLightHandlers;
-import net.minecraft.world.entity.Entity;
+import dev.lambdaurora.lambdynlights.engine.DynamicLightSource;
+import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.item.PrimedTnt;
 import net.minecraft.world.level.Level;
@@ -25,60 +24,40 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(PrimedTnt.class)
-public abstract class PrimedTntEntityMixin extends Entity implements DynamicLightSource {
+public abstract class PrimedTntEntityMixin extends EntityMixin implements DynamicLightSource {
 	@Shadow
 	public abstract int getFuse();
 
 	@Unique
 	private int startFuseTimer = 80;
-	@Unique
-	private int lambdynlights$luminance;
-
-	public PrimedTntEntityMixin(EntityType<?> type, Level level) {
-		super(type, level);
-	}
 
 	@Inject(method = "<init>(Lnet/minecraft/world/entity/EntityType;Lnet/minecraft/world/level/Level;)V", at = @At("TAIL"))
 	private void onNew(EntityType<? extends PrimedTnt> type, Level level, CallbackInfo ci) {
 		this.startFuseTimer = this.getFuse();
 	}
 
-	@Inject(method = "tick", at = @At("TAIL"))
-	private void onTick(CallbackInfo ci) {
-		// We do not want to update the entity on the server.
-		if (this.level().isClientSide()) {
-			if (!LambDynLights.get().config.getTntLightingMode().isEnabled())
-				return;
-
-			if (this.isRemoved()) {
-				this.setDynamicLightEnabled(false);
-			} else {
-				if (!LambDynLights.get().config.getEntitiesLightSource().get() || !DynamicLightHandlers.canLightUp(this))
-					this.resetDynamicLight();
-				else
-					this.dynamicLightTick();
-				LambDynLights.updateTracking(this);
-			}
-		}
+	@Override
+	public boolean isDynamicLightEnabled() {
+		return super.isDynamicLightEnabled() && LambDynLights.get().config.getTntLightingMode().isEnabled();
 	}
 
 	@Override
 	public void dynamicLightTick() {
-		if (this.isOnFire()) {
-			this.lambdynlights$luminance = 15;
-		} else {
-			ExplosiveLightingMode lightingMode = LambDynLights.get().config.getTntLightingMode();
-			if (lightingMode == ExplosiveLightingMode.FANCY) {
-				var fuse = this.getFuse() / this.startFuseTimer;
-				this.lambdynlights$luminance = (int) (-(fuse * fuse) * 10.0) + 10;
-			} else {
-				this.lambdynlights$luminance = 10;
-			}
+		if (!LambDynLights.get().config.getTntLightingMode().isEnabled()) {
+			this.lambdynlights$luminance = 0;
+			return;
 		}
-	}
 
-	@Override
-	public int getLuminance() {
-		return this.lambdynlights$luminance;
+		super.dynamicLightTick();
+
+		ExplosiveLightingMode lightingMode = LambDynLights.get().config.getTntLightingMode();
+		int luminance;
+		if (lightingMode == ExplosiveLightingMode.FANCY) {
+			luminance = (int) (-MathHelper.smoothstep((float) this.getFuse() / this.startFuseTimer) * 10.f) + 10;
+		} else {
+			luminance = 10;
+		}
+
+		this.lambdynlights$luminance = Math.max(this.lambdynlights$luminance, luminance);
 	}
 }